+++ src/sc_client/client/__init__.py
@@ -7,6 +7,9 @@ Distributed under the MIT License
 from sc_client.client._api import (
     check_elements,
     connect,
+    is_connected,
+    set_error_handler,
+    set_reconnect_handler,
     create_elements,
     create_elements_by_scs,
     delete_elements,