+++ src/sc_client/client/_api.py
@@ -8,7 +8,7 @@ from __future__ import annotations
 
 from sc_client import session
 from sc_client.constants import common, exceptions
-from sc_client.constants.numeric import SERVER_RECONNECTION_TIME
+from sc_client.constants.numeric import SERVER_RECONNECT_RETRIES, SERVER_RECONNECT_RETRY_DELAY
 from sc_client.constants.sc_types import ScType
 from sc_client.models import (
     ScAddr,
@@ -26,17 +26,29 @@ from sc_client.models import (
 from sc_client.models.sc_construction import ScLinkContentData
 
 
-def connect(url: str, do_reconnect=False, reconnection_time=SERVER_RECONNECTION_TIME) -> None:
-    session.set_connection(url, do_reconnect, reconnection_time)
+def connect(url: str) -> None:
+    session.set_connection(url)
 
 
 def is_connected() -> bool:
-    return session.is_connection_established()
+    return session.is_connected()
 
 
 def disconnect() -> None:
-    session.disable_reconnection()
     session.close_connection()
+
+
+def set_error_handler(callback) -> None:
+    session.set_error_handler(callback)
+
+
+def set_reconnect_handler(**reconnect_kwargs) -> None:
+    session.set_reconnect_handler(
+        reconnect_kwargs.get("reconnect_handler", session.default_reconnect_handler),
+        reconnect_kwargs.get("post_reconnect_handler"),
+        reconnect_kwargs.get("reconnect_retries", SERVER_RECONNECT_RETRIES),
+        reconnect_kwargs.get("reconnect_retry_delay", SERVER_RECONNECT_RETRY_DELAY)
+    )
 
 
 def check_elements(*addrs: ScAddr) -> list[ScType]:
@@ -80,13 +92,13 @@ def resolve_keynodes(*params: ScIdtfResolveParams) -> list[ScAddr]:
 
 
 def template_search(
-    template: ScTemplate | str | ScTemplateIdtf | ScAddr, params: ScTemplateParams = None
+        template: ScTemplate | str | ScTemplateIdtf | ScAddr, params: ScTemplateParams = None
 ) -> list[ScTemplateResult]:
     return session.execute(common.ClientCommand.SEARCH_TEMPLATE, template, params)
 
 
 def template_generate(
-    template: ScTemplate | str | ScTemplateIdtf | ScAddr, params: ScTemplateParams = None
+        template: ScTemplate | str | ScTemplateIdtf | ScAddr, params: ScTemplateParams = None
 ) -> ScTemplateResult:
     return session.execute(common.ClientCommand.GENERATE_TEMPLATE, template, params)