+++ docs/changelog.md
@@ -5,10 +5,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+ - Reconnection retries to sc-server in send message
+ - OnReconnect handler setter
+ - OnError handler setter
+### Fixed
+ - Infinite loop when receiving response to message
+### Removed
+ - Reconnection to sc-server in separate thread by time
 
 ## [0.2.6]
 ### Added
-- Reconnection to sc-server can be enabled in connect params
+ - Reconnection to sc-server can be enabled in connect params
 ### Changed
  - Decompose sc-types